fix: Allow filter expiration to be set to "indefinite"
Mastodon API uses an "empty" `expires_in` value for a filter to mean
"Does not expire" (i.e., indefinite).

This was modelled as a null. Which doesn't work, because Retrofit does
not send name/value pairs in encoded forms if the value is null.

Fix this by making the API type a `String?`, and explicitly using the
empty string when indefinite expiry is used. This has to be converted
back to an Int? in a few places.

See https://github.com/mastodon/documentation/issues/1216#issuecomment-2030222940

+++ app/src/test/java/app/pachli/FilterV1Test.kt
@@ -260,7 +260,7 @@ class FilterV1Test {
         val expiredBySeconds = 3600
         val expiredDate = Date.from(Instant.now().minusSeconds(expiredBySeconds.toLong()))
         val updatedDuration = EditFilterActivity.getSecondsForDurationIndex(-1, null, expiredDate)
-        assert(updatedDuration != null && updatedDuration <= -expiredBySeconds)
+        assert(updatedDuration != null && updatedDuration.toInt() <= -expiredBySeconds)
     }
 
     @Test
@@ -268,7 +268,7 @@ class FilterV1Test {
         val expiresInSeconds = 3600
         val expiredDate = Date.from(Instant.now().plusSeconds(expiresInSeconds.toLong()))
         val updatedDuration = EditFilterActivity.getSecondsForDurationIndex(-1, null, expiredDate)
-        assert(updatedDuration != null && updatedDuration > (expiresInSeconds - 60))
+        assert(updatedDuration != null && updatedDuration.toInt() > (expiresInSeconds - 60))
     }
 
     companion object {
